Refactoring: Create common basebutton and other btn component use this basebutton
