+++ frontend/src/components/Login/LoginKakaoBtn.vue
@@ -1,11 +1,10 @@
 <template>
-  <v-btn v-bind="btnProps" @click="method">
-    <v-icon v-bind="iconProps">{{ icon }}</v-icon>
-    {{ button }}
-  </v-btn>
+  <BaseButton :btnProps="btnProps" :iconProps="iconProps" :icon="icon" :method="method" :button="button" />
 </template>
 
 <script>
+import BaseButton from "@/common/BaseButton.vue";
+
 export default {
   data: () => ({
     btnProps: {
@@ -15,30 +14,9 @@ export default {
       left: true,
     },
   }),
-  props: {
-    method: {
-      type: Function,
-      required: true,
-      validator: function(value) {
-        return value != null;
-      },
-    },
-    icon: {
-      type: String,
-      required: true,
-      validator: function(value) {
-        if (value.indexOf("fas") != -1) return true;
-        if (value.indexOf("mdi") != -1) return true;
-        return false;
-      },
-    },
-    button: {
-      type: String,
-      required: true,
-      validator: function(value) {
-        return value != null;
-      },
-    },
+  props: ["method", "icon", "button"],
+  components: {
+    BaseButton,
   },
 };
 </script>
